Fixed another console bug, had to update listLists task-info with an id
