+++ public/javascripts/app.js
@@ -1,4 +1,4 @@
-import { handleErrors } from "./utils.js";
+import { handleErrors, addTaskInfoListeners } from "./utils.js";
 import { fetchTask, fetchComments, postComment } from "./task-comments.js";
 
 // fetch user's tasks (all)
@@ -25,6 +25,8 @@ export const fetchTasks = async () => {
     `)
 
     tasksListContainer.innerHTML = listName + tasksHtml.join("");
+
+    await addTaskInfoListeners();
 }
 
 // fetch assigned tasks
@@ -46,7 +48,7 @@ const fetchAssignTasks = async () => {
     `
     const tasksHtml = tasks.map(({ id, description, User }) => `
     <div class="assigned-grid">
-    <div class="task-info">
+    <div class="task-info" id=${id}>
         <input type="checkbox" class="task-check-box" id=${id} name=${id}>
         <label for=${id} id=${id} class="task-check-box">${description}</label>
     </div>
@@ -57,6 +59,8 @@ const fetchAssignTasks = async () => {
     `)
 
     assignedTaskContainer.innerHTML = listName + tasksHtml.join("")
+
+    await addTaskInfoListeners();
   }
 
 }
@@ -78,13 +82,15 @@ export const fetchIncompleteTasks = async () => {
   <h2 class="task-list-header"><strong>${user.username}' s</strong> incomplete tasks.</h2>
   `
   const tasksHtml = tasks.map(({ id, description }) => `
-  <div class="task-info">
+  <div class="task-info" id=${id}>
       <input type="checkbox" class="task-check-box" id=${id} name=${id}>
       <label for=${id} id=${id} class="task-check-box">${description}</label>
   </div>
   `)
 
   tasksListContainer.innerHTML = listName + tasksHtml.join("");
+
+  await addTaskInfoListeners();
 }
 
 // fetch user's completed tasks
@@ -104,13 +110,15 @@ const fetchCompletedTasks = async () => {
   <h2 class="task-list-header"><strong>${user.username}' s</strong> completed tasks.</h2>
   `
   const tasksHtml = tasks.map(({ id, description }) => `
-  <div class="task-info">
+  <div class="task-info" id=${id}>
       <input type="checkbox" class="task-check-box" id=${id} name=${id}>
       <label for=${id} id=${id} class="task-check-box">${description}</label>
   </div>
   `)
 
   tasksListContainer.innerHTML = listName + tasksHtml.join("");
+
+  await addTaskInfoListeners();
 }
 
 // fetch user's search query
@@ -133,7 +141,7 @@ const search = async (searchValue) => {
   <h2 class="task-list-header">Search Results</h2>
   `
   const tasksHtml = results.map(({ id, description, Lists }) => `
-  <div class="task-info">
+  <div class="task-info" id=${id}>
       <input type="checkbox" class="task-check-box" id=${id} name=${id}>
       <label for=${id} id=${id} class="task-check-box"><strong>${description}</strong> found in your <strong><i>${Lists[0].title}</i></strong> list.</label>
   </div>
@@ -141,7 +149,7 @@ const search = async (searchValue) => {
 
   tasksListContainer.innerHTML = listName + tasksHtml.join("");
 
-
+  await addTaskInfoListeners();
 
 
 }
@@ -188,7 +196,7 @@ const fetchContactTasks = async (id) => {
   clearAssignedList.innerHTML = ``;
   }
   const tasksHtml = tasks.map(({ id, description }) => `
-  <div>
+  <div id=${id} class="task-info">
       <input type="checkbox" class="task-check-box" id=${id} name=${id}>
       <label for=${id} class="task-check-box" id=${id}>${description}</label>
   </div>
@@ -234,46 +242,6 @@ document.addEventListener("DOMContentLoaded", async () => {
     } catch (e) {
       console.error(e);
     }
-
-    const taskInfoContainer = document.querySelectorAll('.task-info');
-    taskInfoContainer.forEach( (task) => {
-      task.addEventListener('click', async(e) => {
-        e.stopPropagation();
-        const taskId = task.id;
-
-        const editForm = document.querySelector('.edit-form');
-        editForm.hidden = true;
-
-        try {
-          await fetchTask(taskId);
-
-          const createComment = document.querySelector('.create-comment');
-
-          createComment.addEventListener('submit', async (event) => {
-            event.stopPropagation();
-            event.preventDefault();
-            const commentData = new FormData(createComment);
-            const message = commentData.get("message");
-            const taskId = commentData.get("taskId");
-
-            const body = { message };
-
-            postComment(taskId, body);
-
-          })
-
-        } catch (e) {
-          console.error(e);
-        }
-
-        try {
-          await fetchComments(taskId);
-        } catch (e) {
-          console.error(e);
-        }
-
-      })
-    })
   }
 );
 
@@ -484,13 +452,15 @@ const fetchLists = async () => {
         `
 
         const tasksHtml = tasks.map((task) => `
-          <div class="task-info">
+          <div class="task-info" id=${task.Task.id}>
             <input type="checkbox" class="task-check-box" id=${task.Task.id} name=${task.Task.id}>
             <label for=${task.Task.id} id=${task.Task.id} class="task-check-box">${task.Task.description}</label>
           </div>
         `)
 
         tasksContainer.innerHTML = listTitle + tasksHtml.join('');
+
+        await addTaskInfoListeners();
       })
     })
 }
@@ -696,13 +666,15 @@ givenToOthersList.addEventListener('click', async(e) => {
       <h2 class="task-list-header">Tasks Given to Others</h2>
 `
   const tasksHtml = tasks.map(({ id, description }) => `
-  <div class="task-info">
+  <div class="task-info" id=${id}>
       <input type="checkbox" class="task-check-box" id=${id} name=${id}>
       <label for=${id} id=${id} class="task-check-box">${description}</label>
   </div>
   `)
 
   tasksListContainer.innerHTML = listName + tasksHtml.join("");
+
+  await addTaskInfoListeners();
 })
 
 
@@ -724,11 +696,13 @@ givenToMeList.addEventListener('click', async(e) => {
       <h2 class="task-list-header">Tasks Given to Me</h2>
 `
   const tasksHtml = tasks.map(({ id, description }) => `
-  <div class="task-info">
+  <div class="task-info" id=${id}>
       <input type="checkbox" class="task-check-box" id=${id} name=${id}>
       <label for=${id} id=${id} class="task-check-box">${description}</label>
   </div>
   `)
 
   tasksListContainer.innerHTML = listName + tasksHtml.join("");
+
+  await addTaskInfoListeners();
 })
